requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
langchain==0.3.0
langchain-core==0.3.0
langchain-community==0.3.0
langchain-google-genai==2.0.0
langchain-groq==0.2.0
chromadb==0.5.5
pypdf==4.3.1
sentence-transformers==3.1.1
requests==2.32.5
python-multipart==0.0.9
google-generativeai==0.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4